[intentions] fix conversion from `()` to `invoke()` without lambdas
^KTIJ-133 Fixed
^KTIJ-134 Fixed

+++ kotlin-ide/analysis/src/org/jetbrains/kotlin/idea/intentions/OperatorToFunctionIntention.kt
@@ -231,7 +231,7 @@ class OperatorToFunctionIntention : SelfTargetingIntention<KtExpression>(
             val callExpression = transformed.getCalleeExpressionIfAny()?.parent as? KtCallExpression
             if (callExpression != null) {
                 funcLitArgs.forEach { callExpression.add(it) }
-                if (argumentsWithReceiverIfNeeded.isEmpty()) {
+                if (funcLitArgs.isNotEmpty() && argumentsWithReceiverIfNeeded.isEmpty()) {
                     callExpression.valueArgumentList?.delete()
                 }
             }
